requirements.txt
asgiref==3.5.0
Django==4.0.3
sqlparse==0.4.2
djangorestframework==3.12.4
django_storages==1.12.2
django-cors-headers==3.10.0
django-environ==0.4.5
psycopg2-binary>=2.8
pytz